05_VisionAI_불량판별시스템/requirements.txt
numpy
opencv-python
gspread
oauth2client
pydrive2
requests
pyserial
